fix penalty for escape from battle

+++ Character.cs
@@ -15,7 +15,7 @@ namespace EternityRPG
         public double MaxDamage { get; protected set; }
         public int CritChance { get; set; } = 5;
         public int Exp { get; set; }
-        public int Gold { get; set; }
+        public double Gold { get; set; }
 
 
         public double GenerateDamage() => Damage = random.Next((int) MinDamage, (int) MaxDamage);
